binned accumulator clean up firts cut header needs work
