+++ Vectorisation/VecX/dr3.h
@@ -14,6 +14,7 @@
 #include "vec_bool.h"
 #include "vec_view.h"
 #include "transform.h"
+#include "binned_accumulator.h"
 
 #include "target_name_space.h"
 
@@ -590,7 +591,7 @@ typename InstructionTraits<INS_VEC>::FloatType reduceI(const Vec<INS_VEC>& rhs1,
 
 //unroll user accumulator version
 template< template <class> typename ACCUMULATOR_TYPE, template <class> typename VEC_TYPE, typename INS_VEC, typename OP>
-typename InstructionTraits<INS_VEC>::FloatType reduceII(const ACCUMULATOR_TYPE< INS_VEC>& acc,const VEC_TYPE<INS_VEC>& rhs1, OP& oper, typename InstructionTraits<INS_VEC>::FloatType initVal = InstructionTraits<INS_VEC>::nullValue, bool singularInit = true)
+typename InstructionTraits<INS_VEC>::FloatType reduceWithAccumulator(const ACCUMULATOR_TYPE< INS_VEC>& acc,const VEC_TYPE<INS_VEC>& rhs1, OP& oper, typename InstructionTraits<INS_VEC>::FloatType initVal = InstructionTraits<INS_VEC>::nullValue, bool singularInit = true)
 {
 	ignore(initVal);
 	ignore(singularInit);
@@ -615,7 +616,7 @@ template< typename ACCUMULATOR_TYPE, template <class> typename VEC_TYPE, typenam
 typename InstructionTraits<INS_VEC>::FloatType reduce( const VEC_TYPE<INS_VEC>& rhs1, OP& oper, typename InstructionTraits<INS_VEC>::FloatType initVal = InstructionTraits<INS_VEC>::nullValue, bool singularInit = true)
 {
 	ACCUMULATOR_TYPE Bin;
-	return reduceII(Bin, rhs1, oper);
+	return reduceWithAccumulator(Bin, rhs1, oper);
 };
